refactor: 가게 단일 조회(detail)디렉토리 수정
