+++ src/main/java/com/project/nupibe/domain/store/dto/response/StoreResponseDTO.java
@@ -6,6 +6,16 @@ import java.util.List;
 
 public class StoreResponseDTO {
     @Builder
+    public record StoreDetailResponseDTO (
+            Long id,
+            String name,
+            String content,
+            String image,
+            String category,
+            Integer like_num,
+            Integer bookmark_num){
+    }
+    @Builder
     public record StorePreviewDTO(
             Long id,
             String image,
