--- src/main/java/com/project/nupibe/global/apiPayload/exception/handler/StoreHandler.java
@@ -1,8 +1,0 @@
-package com.project.nupibe.global.apiPayload.exception.handler;
-
-import com.project.nupibe.global.apiPayload.code.BaseErrorCode;
-import com.project.nupibe.global.apiPayload.exception.CustomException;
-
-public class StoreHandler extends CustomException {
-    public StoreHandler(BaseErrorCode errorCode){super(errorCode);}
-}
